// types.ts
export interface Message {
  id: string;
  role: 'user' | 'model';
  text: string;
  isStreaming?: boolean;
  timestamp: Date;
  attachment?: {
    mimeType: string;
    data: string; // Base64 string
  };
  generatedImage?: string; // Base64 string for images generated by AI
}

export enum LoadingState {
  IDLE = 'IDLE',
  LOADING = 'LOADING',
  STREAMING = 'STREAMING',
  GENERATING_IMAGE = 'GENERATING_IMAGE',
  ERROR = 'ERROR'
}

export interface UserProfile {
  name: string;
  email: string; // Login uchun (username o'rniga)
  avatar?: string; // Base64 avatar
}

export interface StoredUser extends UserProfile {
  password: string; 
}